fix(health): 陈旧探测结论不再显示成确定的「不可达」+ 启用 Key 时立即探测
真机现象：一条**禁用** Key 的卡片显示「探测不可达 · 10 天前 · 342ms」，
而那家上游早已恢复、真实转发也能成功。用户据此以为「这个 Key 现在是坏的」，
排查方向完全错了。

根因（日志实证，两点叠加）：
1. 定时健康检查只扫**启用**的 Key → 禁用期间 status 永久冻结在很久以前那次失败上
2. 10 天前 agentrouter 确实在报错（401 / 503 / operation timed out 都出现过），
   之后没有任何东西去刷新那个快照

**先证伪过一个推测**：起初怀疑「探测用对外名导致上游不认」，读码确认
`probe_model()` 取的是映射后的真实名、与转发一致 —— 不是模型名问题。

两处修复：
- `HealthBadge`：down 且 lastChecked 超过 1 小时 → 显示灰色「状态已过期」并提示
  点刷新重新检测，不再给确定结论。**只对 down 做**：乐观的陈旧（up）不会误导用户
  去排查不存在的故障，悲观的陈旧会。阈值 1 小时比定时探测最长间隔宽裕，
  正常启用的 Key 不会被误判。
- `toggle_key`：启用时异步探测一次（不阻塞，探测最长 30s，同步会卡界面半分钟）。
  「刚把它启用」正是最需要知道当下可用性的时刻。

时间单位走 i18n（agoParts 只算数值 + 选 key，措辞交语言包）：组件里硬编码中文
会让英文界面出现「10 天」。浏览器实测 zh「10 天前」/ en「10 d ago」均正确，
且新鲜的 down（30 秒前）仍显示「探测不可达」未被误判为过期。

mock 数据加两条对照 Key（禁用+10天前失败 / 启用+刚失败），让两种措辞在
npm run dev 下能并排看到。顺带清掉本批改动引入的 4 条 clippy 警告。

+++ src-tauri/src/proxy.rs
@@ -1133,6 +1133,10 @@ enum StreamAttempt {
 /// - 先 send() 探状态码：非 2xx 缓冲错误体返回 HttpError（首字节未发，切换安全）；
 ///   2xx 则用 bytes_stream() 逐块转发，content-type 沿用上游真实值（保 text/event-stream）。
 /// - 仅在下游协议与 Key 协议一致时调用（无需跨协议转换），跨协议 SSE 翻译属已知限制。
+// 参数多但每个都是这条转发链必需的运行时上下文（store/分类写日志、key/path/body/model 定位请求、
+// headers 透传客户端身份）。抽成 struct 只是换个地方传同样的东西，还要动全部调用点 ——
+// 与 `Store::append_event_full`、`aggregate::run_member_turns` 同样的取舍。
+#[allow(clippy::too_many_arguments)]
 async fn try_stream_to_key(
     store: &Arc<Store>,
     category: CategoryType,
@@ -1451,6 +1455,8 @@ fn inject_default_effort(
 /// 返回完整 outcome（含发往上游的请求体、响应体、状态），供路由与调用模型日志共用。
 /// 注意：非 2xx 不再直接返回 Err，而是照常返回 outcome（ok=false），
 /// 由调用方决定是否切换——这样失败也能被完整记进调用模型日志。
+// 同 `try_stream_to_key`：参数均为必需的运行时上下文，不为消警告做无收益的重构。
+#[allow(clippy::too_many_arguments)]
 async fn forward_to_key(
     store: &Arc<Store>,
     category: CategoryType,
@@ -2463,7 +2469,7 @@ mod tests {
         );
         // 5 个候选都指向同一个慢上游；每个 Key 自身超时 900ms
         for i in 0..5 {
-            let mut k = key(&format!("k{i}"), i as i32, &slow);
+            let mut k = key(&format!("k{i}"), i, &slow);
             k.params.timeout_ms = Some(900);
             store.upsert_key(k).unwrap();
             store.secrets.write().set(&format!("k{i}"), "x").unwrap();
